# public/api-docs/swagger.yaml
swagger: "2.0"
host: festivals-app.herokuapp.com
basePath: /api
info:
  version: 1.0.0
  title: festivals
...
schemes:
  - https
produces:
  - application/vnd.festivals.v1+json
  - application/vnd.festivals.v1+xml
security:
  - oauth2token:
    - public
paths:
  /festivals:
    get:
      tags:
        - festivals
      summary: Find festivals
      description: Get festivals collection
      operationId: getFestivals
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - festivals:get
      parameters:
        - name: name
          description: name
          required: false
          type: string
          in: query
        - name: startAt
          description: startAt
          required: false
          type: string
          in: query
        - name: updatedAt
          description: updatedAt
          required: false
          type: string
          in: query
        - name: limit
          description: limit
          required: false
          type: integer
          in: query
          default: 100
        - name: offset
          description: offset
          required: false
          type: integer
          in: query
        - name: country
          description: country
          required: false
          type: string
          in: query
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/FestivalsCollectionResponse"
    post:
      tags:
        - festivals
      summary: Create festival
      description: Create festival
      operationId: createFestival
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - festivals:create
      parameters:
        - in: body
          name: festivalRequest
          description: Festival object
          required: true
          schema:
            $ref: "#/definitions/FestivalRequest"
      responses:
        "201":
          description: successful operation
          schema:
            $ref: "#/definitions/FestivalResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
  "/festivals/{id}":
    get:
      tags:
        - festivals
      summary: Get festival
      description: Get festival
      operationId: getFestival
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - festivals:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/FestivalResponse"
        "404":
          description: Festival not found
    delete:
      tags:
        - festivals
      summary: Delete festival
      description: Delete festival
      operationId: deleteFestival
      security:
        - oauth2token:
          - festivals:delete
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
      responses:
        "204":
          description: successful operation
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      tags:
        - festivals
      summary: Update festival
      description: Update festival
      operationId: updateFestival
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - festivals:update
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - in: body
          name: festivalRequest
          description: Festival object
          required: true
          schema:
            $ref: "#/definitions/FestivalRequest"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/FestivalResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Festival not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
...
  "/festivals/{id}/places":
    get:
      tags:
        - places
      summary: Get festival places
      description: Get festival places collection
      operationId: getFestivalPlaces
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - places:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: updatedAt
          description: updatedAt
          required: false
          type: string
          in: query
        - name: parent.id
          description: place parent id
          required: false
          type: string
          in: query
        - name: name
          description: name
          required: false
          type: string
          in: query
        - name: limit
          description: limit
          required: false
          type: integer
          in: query
          default: 100
        - name: offset
          description: offset
          required: false
          type: integer
          in: query
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/PlacesCollectionResponse"
        "404":
          description: Festival not found
    post:
      tags:
        - places
      summary: Create festival place
      description: Create festival place
      operationId: createFestivalPlace
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - places:create
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - in: body
          name: festivalPlaceRequest
          description: Festival place object
          required: true
          schema:
            $ref: "#/definitions/PlaceRequest"
      responses:
        "201":
          description: successful operation
          schema:
            $ref: "#/definitions/PlaceResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
  "/festivals/{id}/places/{place.id}":
    get:
      tags:
        - places
      summary: Get festival place
      description: Get festival place
      operationId: getFestivalPlace
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - places:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: place.id
          in: path
          description: Place id
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/PlaceResponse"
        "404":
          description: Festival not found
    delete:
      tags:
        - places
      summary: Delete festival place
      description: Delete festival place
      operationId: deleteFestivalPlace
      security:
        - oauth2token:
          - places:delete
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: place.id
          in: path
          description: Place id
          required: true
          type: string
      responses:
        "204":
          description: successful operation
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      tags:
        - places
      summary: Update festival place
      description: Update festival place
      operationId: updateFestivalPlace
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - places:update
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: place.id
          in: path
          description: Place id
          required: true
          type: string
        - in: body
          name: festivalPlaceRequest
          description: Festival place object
          required: true
          schema:
            $ref: "#/definitions/PlaceRequest"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/PlaceResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
securityDefinitions:
  oauth2token:
    type: oauth2
    flow: implicit
    authorizationUrl: https://vehicle-history.io/oauth/authorization
    scopes:
      user: user auth scope
definitions:
  CoordinatesRequest:
    properties:
      lat:
        type: number
        format: double
        description: lat
      lon:
        type: number
        format: double
        description: lon
    description: Geo coordinates
    required:
      - lat
      - lon
  CoordinatesResponse:
    properties:
      lat:
        type: number
        format: double
        description: lat
      lon:
        type: number
        format: double
        description: lon
    description: Geo coordinates
    required:
      - lat
      - lon
...
  DurationResponse:
    properties:
      startAt:
        type: string
        description: start at
      finishAt:
        type: string
        description: finish at
      periodMs:
        type: integer
        description: periodMs
    description: Duration response
    required:
      - startAt
      - finishAt
      - periodMs
  DurationRequest:
    properties:
      startAt:
        type: string
        description: start at
      finishAt:
        type: string
        description: finish at
    description: Duration request
    required:
      - startAt
      - finishAt
  LocationResponse:
    properties:
      name:
        type: string
        description: Location name
      state:
        type: string
        description: Location state
      country:
        type: integer
        description: Location country
      street:
        type: string
        description: Location street
      zip:
        type: string
        description: Location zip
      openingTimes:
        type: array
        description: Location opening times
        items:
          $ref: "#/definitions/DurationResponse"
      coordinates:
        $ref: "#/definitions/CoordinatesResponse"
        description: Coordinates
    description: Festival response
    required:
      - name
      - state
      - country
      - street
      - openingTimes
  LocationRequest:
    properties:
      name:
        type: string
        description: Location name
      state:
        type: string
        description: Location state
      country:
        type: integer
        description: Location country
      street:
        type: string
        description: Location street
      zip:
        type: string
        description: Location zip
      openingTimes:
        type: array
        description: Location opening times
        items:
          $ref: "#/definitions/DurationRequest"
      coordinates:
        $ref: "#/definitions/CoordinatesRequest"
        description: Coordinates
    description: Festival response
    required:
      - name
      - state
      - country
      - street
      - openingTimes
  MainImageResponse:
    properties:
      small:
        type: string
        description: Image small url
      medium:
        type: string
        description: Image medium url
      large:
        type: string
        description: Image large url
    description: Festival response
  ImageRequest:
    properties:
      url:
        type: string
        description: Image url
      content:
        type: string
        description: Image content
      order:
        type: integer
        description: Image order
    description: Festival response
    required:
      - order
  PlaceResponse:
    properties:
      id:
        type: string
        description: Festival place id
      name:
        type: string
        description: Festival place name
      parent:
        type: string
        description: Festival place parent
      openingTimes:
        type: array
        description: Festival place opening times
        items:
          $ref: "#/definitions/DurationResponse"
      coordinates:
        $ref: "#/definitions/CoordinatesResponse"
        description: Coordinates
      createdAt:
        type: string
        description: Festival place createdAt
      updatedAt:
        type: string
        description: Festival place updatedAt
    description: Festival response
    required:
      - id
      - name
      - parent
      - openingTimes
      - createdAt
      - updatedAt
  PlaceRequest:
    properties:
      name:
        type: string
        description: Festival place name
      parent:
        type: string
        description: Festival place parent
      openingTimes:
        type: array
        description: Festival place opening times
        items:
          $ref: "#/definitions/Duration"
      coordinates:
        $ref: "#/definitions/CoordinatesRequest"
        description: Coordinates
    description: Festival response
    required:
      - name
      - parent
      - openingTimes
...
  FestivalResponse:
    properties:
      id:
        type: string
        description: Festival id
      name:
        type: string
        description: Festival name
      description:
        type: string
        description: Festival description
      tags:
        type: array
        items:
          type: string
        description: Festival tags
      mainImage:
        $ref: "#/definitions/MainImageResponse"
        description: Festival main image
      duration:
        $ref: "#/definitions/DurationResponse"
        description: Festival duration
      locations:
        type: array
        description: Festival locations
        items:
          $ref: "#/definitions/LocationResponse"
      createdAt:
        type: string
        description: Festival createdAt
      updatedAt:
        type: string
        description: Festival updatedAt
    description: Festival response
    required:
      - id
      - name
      - description
      - duration
      - locations
      - createdAt
      - updatedAt
  FestivalRequest:
    properties:
      name:
        type: string
        description: Festival name
      description:
        type: string
        description: Festival description
      tags:
        type: array
        items:
          type: string
        description: Festival tags
      images:
        type: array
        items:
          $ref: "#/definitions/ImageRequest"
        description: Festival images
      duration:
        $ref: "#/definitions/DurationRequest"
        description: Festival duration
      locations:
        type: array
        description: Festival locations
        items:
          $ref: "#/definitions/LocationRequest"
    description: Festival request
    required:
      - name
      - description
      - duration
      - locations
  FestivalsCollectionResponse:
    properties:
      total:
        type: integer
        description: Total festivals
      festivals:
        type: array
        items:
          $ref: "#/definitions/FestivalResponse"
    description: Festivals collection
    required:
      - total
      - festivals
...
  PlacesCollectionResponse:
    properties:
      total:
        type: integer
        description: Total festival categories
      places:
        type: array
        items:
          $ref: "#/definitions/PlaceResponse"
    description: Festival places collection
    required:
      - total
      - places